+++ src/main/kotlin/AST/ASTUnaryNode.kt
@@ -22,6 +22,7 @@ enum class ASTUnaryTypes {
     IF,
     ELSE,
     ELSE_IF,
+    NEW_LINE,
     NO_PARAM_CALL
 }
 class ASTUnaryNode : ASTTypedNode {
@@ -55,6 +56,7 @@ class ASTUnaryNode : ASTTypedNode {
         ASTUnaryTypes.IF -> "${Compiler.grammar[GrammarToken.KEYWORD_IF]} (${(data as ASTNode).toC()})"
         ASTUnaryTypes.ELSE -> "${Compiler.grammar[GrammarToken.KEYWORD_ELSE]}"
         ASTUnaryTypes.ELSE_IF -> "${Compiler.grammar[GrammarToken.KEYWORD_ELSE]} ${Compiler.grammar[GrammarToken.KEYWORD_IF]} (${(data as ASTNode).toC()})"
+        ASTUnaryTypes.NEW_LINE -> "\n\t${(data as ASTNode).toC().replace("\n","\n\t")}"
         ASTUnaryTypes.NO_PARAM_CALL -> "${(data as ASTNode).toC()}()"
         else -> ""
     }