+++ src/main/kotlin/compiler/Compiler.kt
@@ -5,6 +5,7 @@ import AST.ASTTypedNode
 import czechtina.*
 
 object Compiler {
+    val VERSION = "0.1.4"
     var compilingTo = "C"
     var definedTypes = mutableListOf<String>()
     var definedFunctions = mutableMapOf<String, String>()
@@ -60,6 +61,7 @@ object Compiler {
     }
 
     private fun calcTypePriority(type: String) : Int = when (type) {
+        "none" -> 100
         "pointer" -> 10
         "string" -> 5
         "double" -> 4
@@ -67,6 +69,7 @@ object Compiler {
         "int" -> 2
         "char" -> 1
         "bool" -> 1
+        "void" -> 0
         else -> if (type.contains("array")) 0 else throw Exception("Unhandled Type $type")
     }
 
@@ -74,22 +77,36 @@ object Compiler {
         if (operand == "=")
             return right.getType()
 
-        val leftWeight = calcTypePriority(left.getType())
-        val rightWeight = calcTypePriority(right.getType())
-        val maxW = maxOf(leftWeight, rightWeight)
-        val minW = minOf(leftWeight, rightWeight)
+        try {
+            val leftWeight = calcTypePriority(left.getType())
+            val rightWeight = calcTypePriority(right.getType())
+            val maxW = maxOf(leftWeight, rightWeight)
+            val minW = minOf(leftWeight, rightWeight)
 
-        if (maxW == 10 && minW == 2)
-            throw Exception("cant do this operation with pointer")
+            if (maxW == 10 && minW == 2)
+                throw Exception("cant do this operation with pointer")
 
-        if (operand == "%" && listOf(leftWeight, rightWeight).any { it == 3 || it == 4 })
-            throw Exception("Modulo cant be made with floating point number")
+            if (operand == "%" && listOf(leftWeight, rightWeight).any { it == 3 || it == 4 })
+                throw Exception("Modulo cant be made with floating point number")
 
-        if (Regex(cAndCzechtinaRegex(AllComparation)).matches(operand))
-            return "bool";
+            if (Regex(cAndCzechtinaRegex(AllComparation)).matches(operand))
+                return "bool";
 
-        if (leftWeight > rightWeight)
-            return left.getType()
-        return right.getType()
+            if (leftWeight > rightWeight)
+                return left.getType()
+            return right.getType()
+
+        } catch (e: Exception) {
+            println("------------")
+            println(this)
+            println("operand: $operand")
+            println("left: $left ${left.getType()}")
+            println("right: $right ${right.getType()}")
+            throw Exception("Error in calcBinaryType: ${e.message}")
+        }
+    }
+
+    override fun toString(): String {
+        return "Compiler(compilingTo='$compilingTo', definedTypes=$definedTypes, definedFunctions=$definedFunctions, globalVariables=$globalVariables, localVariable=$localVariable)"
     }
 }